code/report-generator/pyproject.toml
[tool.poetry]
name = "api-scanner-reporter"
version = "0.1.0"
description = "APIScanner Reporter"
authors = ["Cyber Security & Privacy Foundation"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mysql-connector-python = "^8.0.33"
python-docx = "^0.8.11"
defusedxml = "^0.7.1"
pydantic = "^1.10.8"
openapi-spec-validator = "^0.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
